patchLike, patchFollow API 추가

+++ src/user/repository/user_follower.repository.ts
@@ -1,12 +1,8 @@
 import { EntityRepository, getConnection, getRepository, Repository } from "typeorm";
-import { ShowtimeEntity } from "../entity/showtime.entity";
 import { RuntimeException } from "@nestjs/core/errors/exceptions/runtime.exception";
-import { UserEntity } from "../entity/user.entity";
-import { ShowtimeLikeEntity } from "../entity/showtime_like.entity";
-import { CreateShowTimeLikeDto } from "../dto/create.showtimelike.dto";
 import { UserFollowerEntity } from "../entity/user_follower.entity";
 import { CreateUserFollowerDto } from "../dto/create.userfollower.dto";
-import { UserShowtimeEntity } from "../entity/user_showtime.entity";
+import { UserEntity } from "../entity/user.entity";
 
 @EntityRepository(UserFollowerEntity)
 export class UserFollowerRepository extends Repository<UserFollowerEntity> {
@@ -15,19 +11,19 @@ export class UserFollowerRepository extends Repository<UserFollowerEntity> {
    * 쇼타임 팔로우 정보 생성
    * @param createShowTimeDto
    */
-  async patchFollower(createUserFollowerDto: CreateUserFollowerDto) {
+  async patchFollower(userId: number, createUserFollowerDto: CreateUserFollowerDto): Promise<boolean> {
     try {
 
       const userFollowerInfo = await getRepository(UserFollowerEntity)
         .createQueryBuilder('uf')
-        .where('uf.userEntity = :userId', {userId: createUserFollowerDto.userId})
+        .where('uf.userEntity = :userId', {userId: userId})
         .andWhere('uf.followerEntity = :followerId', {userId: createUserFollowerDto.followerId})
         .getOne();
 
       if (!userFollowerInfo) {  // 데이터가 없으므로(팔로우 안했으므로) 팔로우 추가
 
         const userEntity = new UserEntity();
-        userEntity.id = createUserFollowerDto.userId;
+        userEntity.id = userId;
 
         const followerEntity = new UserEntity();
         followerEntity.id = createUserFollowerDto.followerId;
@@ -49,13 +45,15 @@ export class UserFollowerRepository extends Repository<UserFollowerEntity> {
         await getConnection()
           .createQueryBuilder()
           .delete()
-          .from(ShowtimeLikeEntity)
-          .where('id = :userFollowId', {userFollowId: userFollowerInfo.id})
+          .from(UserFollowerEntity)
+          .where('userEntity = :userId', {userId: userId})
+          .andWhere('followerEntity = :followId', {followId: createUserFollowerDto.followerId})
           .execute();
       }
-
+      return true;
     } catch (e) {
       throw new RuntimeException('Server Error. Please try again later.');
+      return false;
     }
   }
